+++ src/components/common/ScrollToBottomListener/ScrollToBottomListener.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 
-import ScrollListener from './ScrollListener';
+import ScrollListener from './ScrollToBottomListener';
 
 describe('ScrollListener ', () => {
     it('should call calback when scrolled to bottom of screen.', () => {
